Work on a asset manager interface

+++ 3DRadSpace/Engine3DRadSpace/Math/Math.cpp
@@ -2,12 +2,12 @@
 
 float Engine3DRadSpace::Math::ToRadians(float degrees)
 {
-    return std::numbers::pi * degrees / 180;
+    return static_cast<float>(std::numbers::pi * degrees / 180);
 }
 
 float Engine3DRadSpace::Math::ToDegrees(float radians)
 {
-    return radians * 180 / std::numbers::pi;
+    return static_cast<float>(radians * 180 / std::numbers::pi);
 }
 
 double Engine3DRadSpace::Math::ToRadians(double degrees)
